requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
pyjwt==2.9.0
python-dotenv==1.0.1
bcrypt==4.2.0
pytest==8.3.3
httpx==0.27.2